bug fix post item

+++ application/controllers/Item.php
@@ -91,7 +91,7 @@ class Item extends CI_Controller {
 		$this->form_validation->set_rules('price', 'Harga', 'required|integer');
 		$this->form_validation->set_rules('item_type', 'Tipe', 'required');
 		$this->form_validation->set_rules('quantity_avalaible', 'Jumlah Stok', 'integer');
-		$this->form_validation->set_rules('unit_weight', 'Berat', 'decimal');
+		$this->form_validation->set_rules('unit_weight', 'Berat', 'integer');
 		$this->form_validation->set_rules('posted_item_description', 'Deskripsi', 'required');
 		$this->form_validation->set_rules('category_id', 'Kategori', 'required');
 		$this->form_validation->set_rules('brand_id', 'Brand', 'required');
